Effects now use Shapes and Points. Circle uses double for radious, not int, closes #60

+++ src/geom/Circle.java
@@ -5,7 +5,7 @@ import javafx.scene.canvas.GraphicsContext;
 
 @SuppressWarnings("restriction")
 public class Circle extends BaseShape{
-	public int radious;
+	public double radious;
 	
 	public Circle(int radious) {
 		super();
